requirements.txt
pyobjc-framework-ApplicationServices
Pillow
git+https://github.com/boppreh/keyboard.git@master
